openwebrtc.recipe: Fix libopenwebrtc_bridge_jni.so (Android)
libopenwebrtc_bridge_jni.so that it depends on libopenwebrtc_jni.so instead of
bundling it statically. In addition, we now depend on libopenwebrtc_jni.so for
all the common dependencies so we don't include multiple copies of libraries
that cannot be loaded multiple times (such as glib). This allows
openwebrtc_bridge.jar to depend to openwebrtc.jar.

Change requested by rugvip