libvpx: Fix build on Android/x86
Even when passing --target=x86-android-gcc it will build for normal Linux. And
on Linux some code will use rand(), which does not exist as a symbol on
Android, and instead is mapped to lrand48() by an inline function.

We have to #define __ANDROID__ to make it do the right thing.